First commit to python branch
energy and magnetization are output files
Code is a bit messy
Made a change to the way the array is stored periodically. Python lists
automatically pick the last element in the array if the index is -1. I
thought this would make the periodic nature of the array easier to work
with if I didn't copy the first two lines. Because this doesn't work in
reverse, I'm not sure what I did was neater.